pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bert-ts-classifier"
version = "0.1.0"
description = "BERT Traditional Chinese classifier (Mainland vs Taiwan) modularized from notebooks"
readme = "README.md"
authors = [{ name = "Edward Huang" }]
license = { text = "Apache-2.0" }
requires-python = ">=3.10,<3.13"
dependencies = [
  "torch>=2.1",
  "transformers>=4.41",
  "pydantic>=2.6",
  "PyYAML>=6.0",
  "scikit-learn>=1.3",
  "matplotlib>=3.8",
  "typer>=0.12",
  "rich>=13.7",
  "fastapi>=0.110",
  "uvicorn>=0.22",
  "nbformat>=5.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "ruff>=0.6.9",
  "black>=24.3.0",
  "mypy>=1.10.0",
  "types-PyYAML",
  "types-requests",
  "pre-commit>=3.7.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["bert_ts_classifier*"]

[tool.pytest.ini_options]
addopts = "-q --maxfail=1 --disable-warnings --cov=src --cov-report=term-missing:skip-covered"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "PL"]
ignore = ["E501"]
exclude = [
  "**/*.ipynb",
  "notebooks/**",
  "examples/**",
  "src/data/**",
  "scripts/**",
  "Gradio_app.py",
  "focal_loss.py",
  "test_inference_v2.ipynb",
]

[tool.black]
line-length = 100
target-version = ["py310"]
preview = true

[tool.mypy]
python_version = "3.10"
strict = false
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = true
mypy_path = "src"
